kafka-lambda-app/stock_market_analysis/requirements.txt
kafka-python==2.0.2
requests==2.30.0
pandas==1.5.3